Update EventManager docs Centralize constants into constants.js Move constants that need to be conditionally required into their own file, to avoid conditional requires where possible
