+++ src/controls/events/constants.js
@@ -1,12 +1,3 @@
-import {
-  Tap,
-  Press,
-  Pinch,
-  Rotate,
-  Pan,
-  Swipe
-} from 'hammerjs';
-
 /**
  * Only one set of basic input events will be fired by Hammer.js:
  * either pointer, touch, or mouse, depending on system support.
@@ -24,6 +15,27 @@ export const BASIC_EVENT_ALIASES = {
   mousedown: 'pointerdown',
   mousemove: 'pointermove',
   mouseup: 'pointerup'
+};
+
+export const INPUT_EVENT_TYPES = {
+  KEY_EVENTS: [
+    'keydown',
+    'keyup'
+  ],
+  MOUSE_EVENTS: [
+    'mousedown',
+    'mousemove',
+    'mouseup',
+    'mouseleave'
+  ],
+  WHEEL_EVENTS: [
+    // Chrome, Safari
+    'wheel',
+    // IE
+    'mousewheel',
+    // legacy Firefox
+    'DOMMouseScroll'
+  ]
 };
 
 /**
@@ -63,16 +75,6 @@ export const EVENT_RECOGNIZER_MAP = {
   swipeup: 'swipe',
   swipedown: 'swipe'
 };
-
-export const RECOGNIZERS = [
-  [Rotate, {enable: false}],
-  [Pinch, {enable: false}, ['rotate']],
-  [Pan, {threshold: 10, enable: false}],
-  [Swipe, {enable: false}],
-  [Press, {enable: false}],
-  [Tap, {event: 'doubletap', taps: 2, enable: false}],
-  [Tap, {enable: false}]
-];
 
 /**
  * Map gestural events typically provided by browsers
